new test scenarios added for items

+++ app/Http/Controllers/ItemController.php
@@ -6,10 +6,11 @@ use App\Http\Requests\ItemRequest;
 use App\Http\Resources\ItemResource;
 use App\Models\Company;
 use App\Models\Item;
-use Illuminate\Http\JsonResponse;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ItemController extends Controller
 {
+    use AuthorizesRequests;
     public function index(Company $company)
     {
         $items = $company->items()->get();
@@ -27,18 +28,16 @@ class ItemController extends Controller
         return new ItemResource($item->load('company'));
     }
 
-    public function update(ItemRequest $request, Item $item, Company $company)
+    public function update(ItemRequest $request, Company $company, Item $item)
     {
-     
-
         $item->update($request->validated());
         return new ItemResource($item->load('company'));
     }
 
-    public function delete(Item $item, Company $company)
+    public function delete(Company $company, Item $item)
     {
-     
+        $this->authorize('delete', $item);
         $item->delete();
-        return response()->json(['message' => 'Ürün başarıyla silindi.'], 200);
+        return response()->noContent();
     }
 }
